♻️Refactor: mapCommuteDataToCalendar 함수 util폴더 분리
